+++ tests/test_main.py
@@ -16,6 +16,7 @@ def test__parse_args__from():
         format=None,
         style=None,
         latest_tag=False,
+        tag_dir="tags",
     )
     assert parse_args(["from", "git"]).vcs == "git"
     assert parse_args(["from", "mercurial"]).vcs == "mercurial"
@@ -31,6 +32,7 @@ def test__parse_args__from():
     assert parse_args(["from", "any", "--style", "pep440"]).style == "pep440"
     assert parse_args(["from", "any", "--style", "semver"]).style == "semver"
     assert parse_args(["from", "any", "--latest-tag"]).latest_tag is True
+    assert parse_args(["from", "any", "--tag-dir", "foo"]).tag_dir == "foo"
     assert parse_args(["from", "subversion", "--tag-dir", "foo"]).tag_dir == "foo"
 
     with pytest.raises(SystemExit):